Google data and user_id are stored in sesssion
This looks good. I'm able to log in with Google, detect
whether I'm new or not to the site. Create a users table
entry, if needed. Google Profile and user_id is stored
in session so that when I need it for creating a new message,
it will be there.